Add java.lang.reflect.ReflectPermission to the test harness's policy, allowing creation of service proxies.  This enables the test harness to run under later versions of JDK7 and JDK8.
It's a common pattern to use dynamically-created proxies to services.  If the service interface is not public (as is the case with the underlying proxy to things like Reggie, Fiddler, etc, creating smart proxies that implement the public interface, which delegates to a "dumb" remote proxy that implements a package-private interface), then the proxy needs to be created in the same package as the interface.  Older versions of the JDK did this quietly, but later versions require the executing code to have a java.lang.reflect.ReflectPermission("newProxyInPackage.xyz") that corresponds to the interface's package.

This patch adds those ReflectPermissions to the QA harness's policy, allowing the harness to run on later versions of the JDK, both JDK7 and JDK8.  Running on an older version of JDK should still be OK, as the permission is ignored if it isn't needed.


git-svn-id: https://svn.apache.org/repos/asf/river/jtsk/branches/2.2@1716512 13f79535-47bb-0310-9956-ffa450edef68