rat_reports.sh was updated to reference latest version of Apache RAT (0.11).
deploy_river.groovy now has '-install' and '-dryrun' options, as well as '-help'.  It's
also written more idiomatically groovy-ish, and doesn't include a version number for the uploads, since
that's in the pom files.  As a result, the same file can be used in the 2.2 branch and in the trunk, and 
it won't need to be updated as the versions increment (just the pom files need to be rev'd).

git-svn-id: https://svn.apache.org/repos/asf/river/jtsk/branches/2.2@1724357 13f79535-47bb-0310-9956-ffa450edef68